refactor: implement Sandi Metz code review feedback for ConnectionErrorHandler
- Extract useConnectionStatus custom hook following SRP
- Create focused components: RetryButton, StatusChip, ErrorDetails
- Reduce ConnectionErrorHandler from 383 to 158 lines
- Improve maintainability through composition over monolithic design
- Maintain all existing functionality while improving code organization

Addresses PR #59 Sandi Metz code review comments:
1. ✅ Single Responsibility Principle compliance
2. ✅ Component size reduction and method complexity improvement
3. ✅ Custom hook extraction for business logic separation
